Update transform and load task
The calc_load_daily_price_other_ccy function now checks how old the
most recent row in the table that stores prices in a different currency
is, and pulls from the source tables only those rows and columns that
are necessary to complete the data in that table.
In addition, in the mentioned table, a primary key has been added to
the "date" column.